+++ packages/manager/modules/cloud-styles/src/index.js
@@ -1,0 +1,1 @@
+import './cloud.less';