+++ packages/manager/modules/cloud-universe-components/rollup.config.js
@@ -1,0 +1,25 @@
+import path from 'path';
+import rollupConfig from '@ovh-ux/component-rollup-config';
+
+const config = rollupConfig({
+  input: './src/index.js',
+}, {
+  lessTildeImporter: {
+    paths: [
+      path.resolve(__dirname, 'node_modules'),
+      path.resolve(__dirname, '../../../../node_modules'),
+    ],
+  },
+});
+
+export default [
+  config.cjs(),
+  config.umd({
+    output: {
+      globals: {
+        angular: 'angular',
+        d3: 'd3',
+      },
+    },
+  }),
+];