+++ packages/manager/modules/cloud-universe-components/src/autoselect/index.js
@@ -1,0 +1,11 @@
+import angular from 'angular';
+
+import directive from './directive';
+
+const moduleName = 'cucAutoselect';
+
+angular
+  .module(moduleName, [])
+  .directive('cucAutoselect', directive);
+
+export default moduleName;