+++ packages/manager/modules/cloud-universe-components/src/cui/page/header/component.js
@@ -1,0 +1,11 @@
+import template from './template.html';
+
+export default {
+  template,
+  transclude: true,
+  bindings: {
+    titleText: '<',
+    subtitleText: '<',
+    guides: '<',
+  },
+};