+++ packages/manager/modules/cloud-universe-components/src/helper/url/service.js
@@ -1,0 +1,8 @@
+import find from 'lodash/find';
+
+export default class CucUrlHelper {
+  static findUrl(item, type) {
+    const urlObj = find(item.urls, url => url.type === type);
+    return urlObj && urlObj.address ? urlObj.address : '';
+  }
+}