+++ packages/manager/modules/cloud-universe-components/src/index.js
@@ -1,0 +1,53 @@
+import angular from 'angular';
+
+import cucAutofocus from './autofocus';
+import cucAutoSelect from './autoselect';
+import cucClickEnterOnKeypress from './clickEnterOnKeypress';
+import cucConfig from './config';
+import cucConsumption from './consumption';
+import cucCurrency from './currency';
+import cucFeatureAvailability from './featureAvailability';
+import cucHelper from './helper';
+import cucHighlightedElement from './highlightedElement';
+import cucMessage from './message';
+import cucMonitoring from './monitoring';
+import cucNavigation from './navigation';
+import cucOrderedHash from './orderedHash';
+import cucPoll from './poll';
+import cucRegion from './region';
+import cucProducts from './products';
+import cucSmoothScrollHere from './smoothScrollHere';
+import cucSpaceMeter from './space-meter';
+import cucUserPref from './user-pref';
+import cucVrack from './vrack';
+import cui from './cui';
+
+const moduleName = 'ngOvhCloudUniverseComponents';
+
+angular
+  .module(moduleName, [
+    cui,
+    cucAutofocus,
+    cucAutoSelect,
+    cucClickEnterOnKeypress,
+    cucConfig,
+    cucConsumption,
+    cucCurrency,
+    cucFeatureAvailability,
+    cucHelper,
+    cucHighlightedElement,
+    cucMessage,
+    cucMonitoring,
+    cucNavigation,
+    cucOrderedHash,
+    cucRegion,
+    cucPoll,
+    cucProducts,
+    cucSmoothScrollHere,
+    cucSpaceMeter,
+    cucUserPref,
+    cucVrack,
+    cui,
+  ]);
+
+export default moduleName;