+++ packages/manager/modules/cloud-universe-components/src/message/index.js
@@ -1,0 +1,14 @@
+import angular from 'angular';
+import '@uirouter/angularjs';
+
+import service from './service';
+
+const moduleName = 'cucMessage';
+
+angular
+  .module(moduleName, [
+    'ui.router',
+  ])
+  .service('CucCloudMessage', service);
+
+export default moduleName;