+++ packages/manager/modules/cloud-universe-components/src/space-meter/legend/index.js
@@ -1,0 +1,16 @@
+import angular from 'angular';
+import 'angular-translate';
+
+import component from './component';
+
+import './index.less';
+
+const moduleName = 'cucCloudSpaceMeterLegend';
+
+angular
+  .module(moduleName, [
+    'pascalprecht.translate',
+  ])
+  .component('cucSpaceMeterLegend', component);
+
+export default moduleName;