+++ packages/manager/modules/core/src/index.js
@@ -15,7 +15,8 @@ import 'angular-dynamic-locale';
 import 'angular-translate';
 import 'angular-translate-loader-pluggable';
 
-import 'ovh-angular-apiv7';
+import '@ovh-ux/ng-ovh-swimming-poll';
+import '@ovh-ux/ng-ovh-api-wrappers';
 import 'ovh-api-services';
 
 import translateServiceProvider from './translate/translate.service';
@@ -29,6 +30,8 @@ const moduleName = 'ovhManagerCore';
 
 angular
   .module(moduleName, [
+    'ngOvhSwimmingPoll',
+    'ngOvhApiWrappers',
     'ngSanitize',
     'angular-translate-loader-pluggable',
     'ovh-api-services',