+++ packages/manager/apps/telecom-task/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [3.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-telecom-task-app@2.0.0...@ovh-ux/manager-telecom-task-app@3.0.0) (2019-03-13)
+
+
+### Build System
+
+* **deps:** upgrade dependencies ([#252](https://github.com/ovh-ux/manager/issues/252)) ([f87f7b7](https://github.com/ovh-ux/manager/commit/f87f7b7))
+
+
+### BREAKING CHANGES
+
+* **deps:** replace both `@ovh-ux/ng-ovh-apiv7` and `ovh-angular-swimming-poll` by `@ovh-ux/ng-ovh-api-wrappers` and `@ovh-ux/ng-ovh-swimming-poll`
+
+
+
+# [2.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-telecom-task-app@1.0.0...@ovh-ux/manager-telecom-task-app@2.0.0) (2019-02-26)
+
+
+### Build System
+
+* **deps:** upgrade ng-ovh-apiv7 to v2.0.0 ([ac6ac62](https://github.com/ovh-ux/manager/commit/ac6ac62))
+
+
+### BREAKING CHANGES
+
+* **deps:** replace `ovh-angular-apiv7` by `@ovh-ux/ng-ovh-apiv7`
+
+
+
 # [1.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-telecom-task-app@0.1.1...@ovh-ux/manager-telecom-task-app@1.0.0) (2019-01-29)