+++ packages/manager/modules/overthebox/CHANGELOG.md
@@ -1,3 +1,27 @@
+# [3.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-overthebox@2.0.1...@ovh-ux/manager-overthebox@3.0.0) (2019-03-13)
+
+
+### Build System
+
+* **deps:** upgrade dependencies ([#252](https://github.com/ovh-ux/manager/issues/252)) ([f87f7b7](https://github.com/ovh-ux/manager/commit/f87f7b7))
+
+
+### BREAKING CHANGES
+
+* **deps:** replace both `@ovh-ux/ng-ovh-apiv7` and `ovh-angular-swimming-poll` by `@ovh-ux/ng-ovh-api-wrappers` and `@ovh-ux/ng-ovh-swimming-poll`
+
+
+
+## [2.0.1](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-overthebox@2.0.0...@ovh-ux/manager-overthebox@2.0.1) (2019-02-28)
+
+
+### Bug Fixes
+
+* **i18n:** add missing translations ([1c99f17](https://github.com/ovh-ux/manager/commit/1c99f17))
+* **i18n:** add missing translations ([701d753](https://github.com/ovh-ux/manager/commit/701d753))
+
+
+
 # [2.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-overthebox@1.1.5...@ovh-ux/manager-overthebox@2.0.0) (2019-01-29)